+++ cli/commands/asset/delete.go
@@ -1,0 +1,52 @@
+package asset
+
+import (
+	"errors"
+	"fmt"
+
+	corev2 "github.com/sensu/sensu-go/api/core/v2"
+	"github.com/sensu/sensu-go/cli"
+	"github.com/sensu/sensu-go/cli/commands/helpers"
+	"github.com/spf13/cobra"
+)
+
+// DeleteCommand adds a command that allows user to delete assets
+func DeleteCommand(cli *cli.SensuCli) *cobra.Command {
+	cmd := &cobra.Command{
+		Use:          "delete [NAME]",
+		Short:        "delete assets",
+		SilenceUsage: true,
+		RunE: func(cmd *cobra.Command, args []string) error {
+			// If no name is present print out usage
+			if len(args) != 1 {
+				_ = cmd.Help()
+				return errors.New("invalid argument(s) received")
+			}
+
+			name := args[0]
+			if skipConfirm, _ := cmd.Flags().GetBool("skip-confirm"); !skipConfirm {
+				if confirmed := helpers.ConfirmDelete(name); !confirmed {
+					fmt.Fprintln(cmd.OutOrStdout(), "Canceled")
+					return nil
+				}
+			}
+
+			asset := &corev2.Asset{
+				ObjectMeta: corev2.ObjectMeta{
+					Name:      name,
+					Namespace: cli.Config.Namespace(),
+				},
+			}
+			if err := cli.Client.Delete(asset.URIPath()); err != nil {
+				return err
+			}
+
+			_, err := fmt.Fprintln(cmd.OutOrStdout(), "Deleted")
+			return err
+		},
+	}
+
+	cmd.Flags().Bool("skip-confirm", false, "skip interactive confirmation prompt")
+
+	return cmd
+}